Refactor GUI related classes into new package
