+++ data/savedata.csv
@@ -1,3 +1,3 @@
 T, 0, NUS Interfaith RAS
-D, 0, CS2103 iP, 2025-02-21
+D, 1, CS2103 iP, 2025-02-21
 E, 0, NUS Career Fair, 2025-02-18, 2025-02-20
