+++ src/main/java/joni/Launcher.java
@@ -1,6 +1,7 @@
 package joni;
 
 import javafx.application.Application;
+import joni.gui.Main;
 
 /**
  * Serves as an entrypoint to launch the application
