+++ src/main/java/joni/gui/DialogBox.java
@@ -1,4 +1,4 @@
-package joni;
+package joni.gui;
 
 import java.io.IOException;
 import java.util.Collections;
